% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{banxico_series}
\alias{banxico_series}
\title{Returns data series from BANXICO}
\usage{
banxico_series(series, metadata = FALSE, verbose = FALSE)
}
\arguments{
\item{series}{Series ID}

\item{metadata}{If TRUE returns list with metadata information}

\item{verbose}{If TRUE prints steps while executing}
}
\value{
data.frame
}
\description{
Returns data.frame with BANXICO data series
}
\examples{
# Bank of Mexico international reserves
\dontrun{
reserves <- banxico_series("SF110168")
}

}
\author{
Eduardo Flores
}

