% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BanxicoCatalog}
\alias{BanxicoCatalog}
\title{Catalog of some ID's}
\format{A data frame with 133 rows and 9 variables:
\describe{
  \item{PARENT}{Parent or theme of data series}
  \item{LEVEL_1}{Level one category}
  \item{LEVEL_2}{Level two category}
  \item{LEVEL_3}{Level three category}
  \item{LEVEL_4}{Level four category}
  \item{LEVEL_5}{Level five category}
  \item{LEVEL_6}{Level six category}
  \item{FREQUENCY}{Frequency of data (i.e. monthly)}
  \item{ID}{ID to pass to banxico series}
}}
\usage{
BanxicoCatalog
}
\description{
A dataset containing the id's for some Banxico series.
}
\keyword{datasets}

