% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{compact_banxico_series}
\alias{compact_banxico_series}
\title{Compacts data and metadata into a data.frame}
\usage{
compact_banxico_series(series)
}
\arguments{
\item{series}{series ID}
}
\description{
Returns data.frame with metadata and data from \code{banxico_series()} in data.frame form. Each metadata data is replicated in its corresponding column.
}
\examples{
\dontrun{
df <- compact_banxico_series("SF110168")
}

}
\author{
Eduardo Flores
}

