% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{helpers}
\alias{banxico_parsemeta}
\alias{banxico_parsetrim}
\alias{helpers}
\title{Helper functions for banxico series}
\usage{
banxico_parsetrim(string, trim_begin = TRUE)

banxico_parsemeta(slist, lookfor, exclude = FALSE)
}
\description{
See details
}
\details{
\code{banxico_parsetrim} translates banxico trimesters to dates. 
\code{banxico_parsemeta} extracts metadata from banxico iqy call.
}

