% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{helpers}
\alias{helpers}
\alias{banxico_parsetrim}
\alias{banxico_parsemeta}
\title{Helper functions for banxico series}
\usage{
banxico_parsetrim(string, trim_begin = TRUE)

banxico_parsemeta(slist, lookfor, exclude = FALSE)
}
\arguments{
\item{string}{x}

\item{trim_begin}{y}

\item{slist}{z}

\item{lookfor}{m}

\item{exclude}{d}
}
\description{
See details
}
\details{
\code{banxico_parsetrim} translates banxico trimesters to dates.
\code{banxico_parsemeta} extracts metadata from banxico iqy call.
}
\examples{
# trimester
string <- "Jan-Mar 2015"
trim <- banxico_parsetrim(string)
}
